% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCandCutoff_parallel.R
\name{OCC.Table}
\alias{OCC.Table}
\title{OCC.Table}
\usage{
OCC.Table(
  N.sim,
  N.max,
  mu.trt,
  Sigma.trt,
  mu.ctrl,
  Sigma.ctrl,
  cens_upper,
  design,
  cohort,
  recruit.int,
  m0,
  L0,
  v0,
  S0,
  time_max,
  eta,
  lambda,
  thin_MCMC,
  Niter
)
}
\arguments{
\item{N.sim}{Number of simulations.}

\item{N.max}{Maximum number of patients to enroll.}

\item{mu.trt}{True mean for treatment arm (log).}

\item{Sigma.trt}{True variance for treatment arm (log).}

\item{mu.ctrl}{True mean for control arm (log).}

\item{Sigma.ctrl}{True variance for control arm (log).}

\item{cens_upper}{Upper limit for the censoring time.}

\item{design}{A numeric value indicating the type of design. 1 = proposed design, 2 = time to recurrence design, 3= time to death design, 4 = time to first event design.}

\item{cohort}{Interim cohort.}

\item{recruit.int}{= Recruitment interval.}

\item{m0}{Prior mean for mu.trt/mu.ctrl.}

\item{L0}{Prior variance for mu.trt/mu.ctrl.}

\item{v0}{For the proposed design, this is the prior degrees of freedom for Sigma.trt/Sigma.ctrl. For the traditional designs, v0/2 is the prior shape for Sigma.trt/Sigma.ctrl.}

\item{S0}{For the proposed design, this is the prior scale matrix for Sigma.trt/Sigma.ctrl. For the traditional designs, v0*S0/2 is the prior scale for Sigma.trt/Sigma.ctrl.}

\item{time_max}{The upper limit for the recurrence and death time sampled from truncated normal. This will set the upper limit to to time_max rather than Inf.}

\item{eta}{A pre-specified lower bound of acceptable performance based on historical information.}

\item{lambda}{Cutoff parameter.}

\item{thin_MCMC}{Thinning degree.}

\item{Niter}{Number of iterations for gibbs sampler.}
}
\value{
A list with the following components:\tabular{ll}{
   \code{mysim1} \tab A numeric vector. mysim1 = c(PRN, PEN,  EN), where PRN is the average percentage of trials that are not stopped, PEN is the average percentage of early termination, and EN is the average number of patients.  \cr
   \tab \cr
   \code{mu} \tab A numeric vector. mu = c(mu.trt,mu.ctrl).  \cr
   \tab \cr
   \code{Sigma.trt} \tab True covariance matrix for treatment arm (log). \cr
   \tab \cr
  \code{Sigma.ctrl} \tab True covariance matrix for control arm (log). \cr
   \tab \cr
   \code{lambda} \tab Cutoff parameter. \cr
}
}
\description{
This is the main function used to generate simulation data, and evaluate the operating characteristics and cutoff value based on the simulation data.
Note that mclapply is used for parallel programming, which works on Unix-style operating systems.
}
\examples{
\dontrun{
OCC.Table(N.sim = 2, N.max = 100,mu.trt = c(log(2), log(5)), 
Sigma.trt = matrix(c(1, 0.5, 0.5, 1), ncol=2),
mu.ctrl= c(log(2), log(5)),
Sigma.ctrl = matrix(c(1, 0.5, 0.5, 1), ncol=2), 
cens_upper = 20,
design = 1, 
cohort = c(40,60,80,100),
recruit.int  = 0.25,
m0 = c(0,0),L0 = diag(10^6, 2), 
v0 = 4,S0 = diag(10^(-6), 2),
time_max = 20,eta = 1,lambda = 0.25,
 thin_MCMC = 5,Niter = 100000)
}
}
