% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: Performing Bayesian Monitoring Using the Win Ratio Approach.}
\description{
This package performs bayesian monitoring Using the win ratio approach for the manuscript "The Win Ratio Approach in Bayesian Monitoring for Two-Arm Phase II Clinical Trial Designs with Multiple Time-to-Event Endpoints".
}
\author{
\strong{Maintainer}: Xinran Huang \email{huangxrtc@gmail.com} (\href{https://orcid.org/0000-0002-8126-9735}{ORCID})

}
\keyword{internal}
